Extract temporary directory context for unit tests
